Generify StyleClassFinder
