+++ repast.simphony.visualization/src/repast/simphony/visualization/gui/EdgeStyleStep.java
@@ -381,7 +381,7 @@ public class EdgeStyleStep extends PluginWizardStep {
 		List<String> foundStyleClasses = styleCache.get(styleInterface);
 
 		if (foundStyleClasses == null){
-			foundStyleClasses = StyleClassFinder.getAvailableStyles(model.getContext(), styleInterface);
+			foundStyleClasses = ClassFinder.getFoundClasses(model.getContext(), styleInterface);
 
 			if (defaultStyle != null){
 				foundStyleClasses.add(defaultStyle);
