+++ repast.simphony.visualization/src/repast/simphony/visualization/gui/StyleClassStep.java
@@ -380,7 +380,7 @@ public class StyleClassStep extends StyleStep {
 		List<String> foundStyleClasses = styleCache.get(styleInterface);
 
 		if (foundStyleClasses == null){
-			foundStyleClasses = StyleClassFinder.getAvailableStyles(model.getContext(), styleInterface);
+			foundStyleClasses = ClassFinder.getFoundClasses(model.getContext(), styleInterface);
 
 			if (defaultStyle != null){
 				foundStyleClasses.add(defaultStyle);
